Factor out a "Paged" class from the "CompletablePaged" class
